+++ README.md
@@ -1,3 +1,8 @@
-# conda-pip
+# conda-pypi
 
-`pip` wrappers to make it work with `conda` in a better way.
+Better PyPI interoperability for the conda ecosystem.
+
+Includes:
+
+- `conda pip`: A subcommand that wraps `pip` to make it work with `conda` in a better way.
+- Adds `EXTERNALLY-MANAGED` to your environments.
