+++ conda_pypi/dependencies/pip.py
