--- pixi.toml
@@ -1,22 +1,0 @@
-[project]
-name = "conda-pip"
-version = "0.1.0"
-description = "`pip` wrappers to make it work with `conda` in a better way."
-authors = ["jaimergp <jaimergp@users.noreply.github.com>"]
-channels = ["conda-forge"]
-platforms = ["linux-64", "osx-64", "osx-arm64", "win-64"]
-
-[tasks]
-dev = 'python -mpip install -e. --no-deps && cp conda_pip/data/EXTERNALLY-MANAGED $(python -c "import sysconfig; print(sysconfig.get_paths()[\"stdlib\"])")'
-test = "pytest -vvv"
-build = "conda build recipe"
-
-[dependencies]
-python = "3.10.* *cpython*"
-conda = ">=23.9"
-pip = "*"
-grayskull = "*"
-pytest = "7.4.3.*"
-fmt = "!=10.2.0"
-pytest-mock = "3.12.0.*"
-conda-build = "*"
