+++ pyproject.toml
@@ -3,8 +3,8 @@ requires = ["hatchling", "hatch-vcs"]
 build-backend = "hatchling.build"
 
 [project]
-name = "conda-pip"
-description = "`pip` wrappers to make it work with `conda` in a better way."
+name = "conda-pypi"
+description = "Better PyPI interoperability for the conda ecosystem."
 readme = "README.md"
 authors = [
   {name = "Jaime Rodríguez-Guerra", email = "jrodriguez@quansight.com"}
@@ -33,10 +33,29 @@ dynamic = [
 ]
 
 [project.urls]
-homepage = "https://github.com/jaimergp/conda-pip"
+homepage = "https://github.com/jaimergp/conda-pypi"
 
 [project.entry-points.conda]
-conda-pip = "conda_pip.plugin"
+conda-pypi = "conda_pypi.plugin"
+
+[tool.pixi.project]
+channels = ["conda-forge"]
+platforms = ["linux-64", "osx-64", "osx-arm64", "win-64"]
+
+[tool.pixi.dependencies]
+python = "3.10.* *cpython*"
+conda = ">=23.9"
+pip = "*"
+grayskull = "*"
+pytest = "7.4.3.*"
+fmt = "!=10.2.0"
+pytest-mock = "3.12.0.*"
+conda-build = "*"
+
+[tool.pixi.tasks]
+dev = 'python -mpip install -e. --no-deps && cp conda_pypi/data/EXTERNALLY-MANAGED $(python -c "import sysconfig; print(sysconfig.get_paths()[\"stdlib\"])")'
+test = "pytest -vvv"
+build = "conda build recipe"
 
 [tool.hatch.version]
 source = "vcs"
